Changed question marks from double to integer

+++ app/src/main/java/com/sunway/averychoke/studywifidirect3/controller/teacher_class/quiz/CreateQuestionAdapter.java
@@ -168,7 +168,7 @@ public class CreateQuestionAdapter extends RecyclerView.Adapter<CreateQuestionAd
                 public void afterTextChanged(Editable s) {
                     if (!TextUtils.isEmpty(s.toString().trim())) {
                         try {
-                            mQuestion.setTotalMarks(Double.parseDouble(s.toString()));
+                            mQuestion.setTotalMarks(Integer.parseInt(s.toString()));
                         } catch (NumberFormatException e) {
                             e.printStackTrace();
                         }
